충돌 해결

+++ src/main/java/com/example/peopoolbe/member/service/MemberService.java
@@ -19,6 +19,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.security.Principal;
+import java.util.Arrays;
 
 @Service
 @RequiredArgsConstructor
@@ -74,7 +75,6 @@ public class MemberService {
 
     public void addRefreshTokenInCookie(TokenResDto tokenResDto, HttpServletResponse response) {
         Cookie refreshCookie = new Cookie("refreshToken", tokenResDto.refreshToken());
-        refreshCookie.setPath("/");
         refreshCookie.setPath("/");
         refreshCookie.setHttpOnly(true);
         refreshCookie.setSecure(true);
@@ -150,6 +150,9 @@ public class MemberService {
     public void logout(HttpServletRequest request, HttpServletResponse response) {
         String refreshToken = null;
         Cookie[] cookies = request.getCookies();
+        for(Cookie cookie : cookies) {
+            System.out.println(cookie.getName() + " + " + cookie.getValue());
+        }
 
         if (cookies != null) {
             for (Cookie cookie : cookies) {
@@ -159,11 +162,17 @@ public class MemberService {
                 }
             }
         }
-        refreshTokenRepository.deleteByRefreshToken(refreshToken);
 
-        Cookie refreshCookie = new Cookie("refreshToken", refreshToken);
+        if(refreshToken != null) {
+            refreshTokenRepository.deleteByRefreshToken(refreshToken);
+        }
+
+        Cookie refreshCookie = new Cookie("refreshToken", "");
         refreshCookie.setPath("/");
         refreshCookie.setMaxAge(0);
+        refreshCookie.setHttpOnly(true);
+        refreshCookie.setSecure(true);
+        refreshCookie.setAttribute("SameSite", "None");
         response.addCookie(refreshCookie);
     }
 }
